Deprecated: Add codelist data to downloaded data

Deprecated, use sp_add_codelist() instead.

add_codelist(
  data,
  codelist = NULL,
  period_column = .data$period_vykaz,
  redownload = FALSE,
  dest_dir = NULL
)

Arguments

data
a data frame returned by sp_get_table().

codelist
The codelist to add. Either a character vector of length one (see sp_tables for possible values), or a data frame returned by sp_get_codelist().

period_column
Unquoted column name of column identifying the data period in data. Leave to default if you have not changed the data object returned by sp_get_table().

redownload
Redownload even if file has already been downloaded? Defaults to FALSE.

dest_dir
character. Directory in which downloaded files will be stored. If left unset, will use the statnipokladna.dest_dir option if the option is set, and tempdir() otherwise. Will be created if it does not exist.

Value

A data frame of same length as data, with added columns from codelist. See Details.

Details

See also