Get dataset documentation

Downloads XLS file with dataset documentation, or opens link to this file in browser.

sp_get_dataset_doc(dataset_id, dest_dir = NULL, download = TRUE)

Arguments

dataset_id

dataset ID. See sp_datasets.

dest_dir

character. Directory in which downloaded files will be stored. If left unset, will use the statnipokladna.dest_dir option if the option is set, and tempdir() otherwise. Will be created if it does not exist.

download

Whether to download (the default) or open link in browser.

Value

(invisible) path to file if download = TRUE, URL otherwise

See also

Examples

if (FALSE) { # \dontrun{
sp_get_dataset_doc("finm")
} # }