Add codelist data to downloaded data

Joins a provided codelist, or downloads and processes one if necessary, and adds it to the data.

sp_add_codelist(
  data,
  codelist = NULL,
  period_column = .data$period_vykaz,
  by = NULL,
  redownload = FALSE,
  dest_dir = NULL
)

Arguments

data
a data frame returned by sp_get_table().

codelist
The codelist to add. Either a character vector of length one (see sp_tables for possible values), or a data frame returned by sp_get_codelist().

period_column
Unquoted column name of column identifying the data period in data. Leave to default if you have not changed the data object returned by sp_get_table().

by
character. Columns by which to join the codelist. Same form as for dplyr::left_join()``..

redownload
Redownload even if file has already been downloaded? Defaults to FALSE.

dest_dir
character. Directory in which downloaded files will be stored. If left unset, will use the statnipokladna.dest_dir option if the option is set, and tempdir() otherwise. Will be created if it does not exist.

Value

A tibble of same length as data, with added columns from codelist. See Details.

Details

The data argument should be a data frame produced by sp_get_table() If this is true, the period_column argument is not needed. The codelist argument, if a data frame, should be a data frame produced by sp_get_codelist(). Specifically, it assumes it contains the following columns:

start_date, a date

end_date, a date

column with the code, character usually named the same as the codelist

#' You can usually tell which codelist you need from the name of the column whose code you are looking to expand, e.g. the codes in column paragraf can be expanded by codelist paragraf.

The function filters the codelist to obtain a set of entries relevant to the time period of data. If data contains tables for multiple periods, this is handled appropriately. Codelist-originating columns in the resulting data frame are renamed so they do not interfere with joining additional codelists, perhaps in a single pipe call.

Note that some codelists are "secondary" and can only be joined onto other codelists. If a codelist does not join using sp_add_codelis(), store the output of sp_get_codelist() and join it manually using dplyr.

See also

Examples

# \donttest{
sp_get_table("budget-central", 2017) %>%
  sp_add_codelist("polozka") %>%
  sp_add_codelist("paragraf")

#> Warning: Either year or month not set.
#> Using defaults of 2017 and 12.

#> ● Set these values explicitly for reproducibility as the defaults may change in the future
#>   to provide access to the latest data by default.

#> ℹ Building URL for dataset 'misris': FIN 1-12 OSS - Plnění rozpočtu KAP a OSS, '2017-12'

#> ℹ Get the dataset documentation at 'http:/monitor.statnipokladna.cz/data/struktura/misris.xlsx'

#> ✔ Storing downloaded archive in and extracting to '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp6Fs44U/misris/2017/12/'

#> ℹ Set dest_dir for more control over downloaded files.

#> ℹ Reading data...

#> ℹ Transforming data...

#> ℹ Building URL for codelist 'polozka' - Rozpočtová položka

#> ✔ Storing codelist in '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp6Fs44U/'

#> ℹ Set dest_dir for more control over downloaded files.

#> ℹ Processing codelist data

#> ℹ Building URL for codelist 'paragraf' - Paragraf

#> ✔ Storing codelist in '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp6Fs44U/'

#> ℹ Set dest_dir for more control over downloaded files.

#> ℹ Processing codelist data

#> # A tibble: 1,058,136 x 30
#>    per_yr per_m ucjed ico   kapitola finmisto polozka paragraf zdroj ZC_EDS
#>    <chr>  <chr> <chr> <chr> <chr>    <chr>    <chr>   <chr>    <chr> <chr> 
#>  1 2017   01    1000… 0000… 0302     3020001  5194    611100   1100… NA    
#>  2 2017   01    1000… 0000… 0302     3020001  5194    611100   4100… NA    
#>  3 2017   01    1000… 0000… 0302     3020001  5195    611100   1100… NA    
#>  4 2017   01    1000… 0000… 0302     3020001  5196    611100   1100… NA    
#>  5 2017   01    1000… 0000… 0302     3020001  5196    611100   1100… NA    
#>  6 2017   01    1000… 0000… 0302     3020001  5196    611100   4100… NA    
#>  7 2017   01    1000… 0000… 0302     3020001  5342    611100   1100… NA    
#>  8 2017   01    1000… 0000… 0302     3020001  5342    611100   4100… NA    
#>  9 2017   01    1000… 0000… 0302     3020001  5362    611100   1100… NA    
#> 10 2017   01    1000… 0000… 0302     3020001  5362    611100   4100… NA    
#> # … with 1,058,126 more rows, and 20 more variables: ZC_UCRIS <chr>,
#> #   ZC_PVS <chr>, budget_adopted <dbl>, budget_amended <dbl>,
#> #   budget_final <dbl>, ZU_OBLIG <dbl>, budget_spending <dbl>,
#> #   period_vykaz <date>, polozka_start_date <date>, polozka_end_date <date>,
#> #   trida <chr>, seskupeni <chr>, podseskupeni <chr>, polozka_nazev <chr>,
#> #   skupina <chr>, pododdil <chr>, oddil <chr>, paragraf_paragraf_nazev <chr>,
#> #   paragraf_start_date <date>, paragraf_end_date <date>

pol <- sp_get_codelist("paragraf")

#> ℹ Codelist file already in '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp6Fs44U/', not downloading. Set `redownload = TRUE` if needed.

#> ℹ Processing codelist data

par <- sp_get_codelist("polozka")

#> ℹ Codelist file already in '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp6Fs44U/', not downloading. Set `redownload = TRUE` if needed.

#> ℹ Processing codelist data

sp_get_table("budget-central", 2017) %>%
  sp_add_codelist(pol) %>%
  sp_add_codelist(par)

#> Warning: Either year or month not set.
#> Using defaults of 2017 and 12.

#> ● Set these values explicitly for reproducibility as the defaults may change in the future
#>   to provide access to the latest data by default.

#> ℹ Files already in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp6Fs44U/misris/2017/12, not downloading. Set `redownload = TRUE` if needed.

#> ℹ Reading data...

#> ℹ Transforming data...

#> # A tibble: 1,058,136 x 30
#>    per_yr per_m ucjed ico   kapitola finmisto polozka paragraf zdroj ZC_EDS
#>    <chr>  <chr> <chr> <chr> <chr>    <chr>    <chr>   <chr>    <chr> <chr> 
#>  1 2017   01    1000… 0000… 0302     3020001  5194    611100   1100… NA    
#>  2 2017   01    1000… 0000… 0302     3020001  5194    611100   4100… NA    
#>  3 2017   01    1000… 0000… 0302     3020001  5195    611100   1100… NA    
#>  4 2017   01    1000… 0000… 0302     3020001  5196    611100   1100… NA    
#>  5 2017   01    1000… 0000… 0302     3020001  5196    611100   1100… NA    
#>  6 2017   01    1000… 0000… 0302     3020001  5196    611100   4100… NA    
#>  7 2017   01    1000… 0000… 0302     3020001  5342    611100   1100… NA    
#>  8 2017   01    1000… 0000… 0302     3020001  5342    611100   4100… NA    
#>  9 2017   01    1000… 0000… 0302     3020001  5362    611100   1100… NA    
#> 10 2017   01    1000… 0000… 0302     3020001  5362    611100   4100… NA    
#> # … with 1,058,126 more rows, and 20 more variables: ZC_UCRIS <chr>,
#> #   ZC_PVS <chr>, budget_adopted <dbl>, budget_amended <dbl>,
#> #   budget_final <dbl>, ZU_OBLIG <dbl>, budget_spending <dbl>,
#> #   period_vykaz <date>, skupina <chr>, pododdil <chr>, oddil <chr>,
#> #   pol_paragraf_nazev <chr>, pol_start_date <date>, pol_end_date <date>,
#> #   par_start_date <date>, par_end_date <date>, trida <chr>, seskupeni <chr>,
#> #   podseskupeni <chr>, par_nazev <chr>

# }